Change the size of a move to 32 bits
Decrease the size of a move integer from 64 bits to 32 bits.
Originally, 39 bits were allocated for the move score, but that
many is unnecessary. There are not very many types of moves
in chess (less than 100) so using the remaining 7 bits of a 32 bit
integer will suffice. Change the moveList to contain moves of only
32 bits and update every function that uses moves.

+++ debug.c
@@ -130,13 +130,13 @@ int checkBoard(const Board* board) {
  * king or the promotion flag and castle flag both being set. This is a
  * debugging function that should only be used inside assert statements.
  * 
- * move:              A 64-bit integer containing all the information about
+ * move:              An integer containing all the information about
  *                    a single move.
  * 
  * return:            True if the move is valid. If the move is not valid, one
  *                    of the internal asserts should fail.
  */
-int validMove(uint64 move) {
+int validMove(int move) {
     int from = move & 0x3F;
     int to = (move >> 6) & 0x3F;
     int captured = (move >> 12) & 0xF;
